requirements.txt
botbuilder-integration-aiohttp>=4.15.0
openai==1.31.1
python-dotenv==1.0.1